# contracts/tools.openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "fields",
          "message"
        ]
      },
      "Capabilities": {
        "properties": {
          "maxInputBytes": {
            "format": "int32",
            "type": "integer"
          },
          "maxOutputBytes": {
            "format": "int32",
            "type": "integer"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array"
          }
        },
        "required": [
          "maxInputBytes",
          "maxOutputBytes",
          "sections"
        ],
        "type": "object"
      },
      "Choice": {
        "properties": {
          "extension": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "mime": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "extension",
          "id",
          "kind",
          "label",
          "mime",
          "note"
        ],
        "type": "object"
      },
      "ExecutionResult": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Failure"
          },
          "filename": {
            "type": "string"
          },
          "inspection": {
            "$ref": "#/components/schemas/Inspection"
          },
          "mime": {
            "type": "string"
          },
          "utility": {
            "$ref": "#/components/schemas/UtilityOutput"
          }
        },
        "type": "object"
      },
      "ExecutionTicket": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "token"
        ],
        "type": "object"
      },
      "ExecutionView": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ExecutionResult"
          },
          "state": {
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "CANCELLED",
              "TIMED_OUT"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "state"
        ],
        "type": "object"
      },
      "Failure": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "Inspection": {
        "properties": {
          "category": {
            "type": "string"
          },
          "details": {
            "additionalProperties": {},
            "type": "object"
          },
          "filename": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/Choice"
            },
            "type": "array"
          },
          "size": {
            "format": "int64",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "category",
          "details",
          "filename",
          "format",
          "outputs",
          "size",
          "warnings"
        ],
        "type": "object"
      },
      "JsonNode": {},
      "JsonRequest": {
        "properties": {
          "input": {
            "maxLength": 100000,
            "minLength": 0,
            "type": "string"
          },
          "mode": {
            "enum": [
              "FORMAT",
              "MINIFY"
            ],
            "type": "string"
          }
        },
        "required": [
          "input",
          "mode"
        ],
        "type": "object"
      },
      "JsonResult": {
        "properties": {
          "mode": {
            "enum": [
              "FORMAT",
              "MINIFY"
            ],
            "type": "string"
          },
          "output": {
            "type": "string"
          },
          "toolId": {
            "type": "string"
          }
        },
        "required": [
          "mode",
          "output",
          "toolId"
        ],
        "type": "object"
      },
      "Section": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "sources",
          "status"
        ],
        "type": "object"
      },
      "Tool": {
        "properties": {
          "id": {
            "type": "string"
          },
          "maxInputLength": {
            "format": "int32",
            "type": "integer"
          },
          "modes": {
            "items": {
              "enum": [
                "FORMAT",
                "MINIFY"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "UtilityOutput": {
        "properties": {
          "output": {
            "type": "string"
          },
          "toolId": {
            "type": "string"
          }
        },
        "required": [
          "output",
          "toolId"
        ],
        "type": "object"
      },
      "UtilityRequest": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "input": {
            "maxLength": 100000,
            "minLength": 0,
            "type": "string"
          },
          "mode": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "input",
          "mode"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/tools": {
      "get": {
        "operationId": "catalog",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Tool"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "tools-controller"
        ]
      }
    },
    "/api/tools/executions/{id}": {
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Execution-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "execution-controller"
        ]
      },
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Execution-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "execution-controller"
        ]
      }
    },
    "/api/tools/executions/{id}/download": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Execution-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "execution-controller"
        ]
      }
    },
    "/api/tools/files/capabilities": {
      "get": {
        "operationId": "capabilities",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "file-controller"
        ]
      }
    },
    "/api/tools/files/convert": {
      "post": {
        "operationId": "convert",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": false,
            "schema": {
              "default": "upload",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sheet",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "width",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "quality",
            "required": false,
            "schema": {
              "default": 90,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "format": "byte",
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "file-controller"
        ]
      }
    },
    "/api/tools/files/executions": {
      "post": {
        "operationId": "file",
        "parameters": [
          {
            "in": "query",
            "name": "operation",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filename",
            "required": false,
            "schema": {
              "default": "upload",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sheet",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "width",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "quality",
            "required": false,
            "schema": {
              "default": 90,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionTicket"
                }
              }
            },
            "description": "Accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "execution-controller"
        ]
      }
    },
    "/api/tools/files/inspect": {
      "post": {
        "operationId": "inspect",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": false,
            "schema": {
              "default": "upload",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Inspection"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "file-controller"
        ]
      }
    },
    "/api/tools/json-formatter": {
      "post": {
        "operationId": "transform",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JsonResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "tools-controller"
        ]
      }
    },
    "/api/tools/utilities": {
      "get": {
        "operationId": "definitions",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JsonNode"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "execution-controller"
        ]
      }
    },
    "/api/tools/{toolId}/executions": {
      "post": {
        "operationId": "utility",
        "parameters": [
          {
            "in": "path",
            "name": "toolId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UtilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionTicket"
                }
              }
            },
            "description": "Accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "tags": [
          "execution-controller"
        ]
      }
    }
  }
}